pyproject.toml
[project]
name = "ray-distributed-computing"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "matplotlib>=3.10.1",
    "numpy>=2.2.3",
    "pandas>=2.2.3",
    "python-dotenv>=1.0.1",
    "ray>=2.43.0",
    "scikit-learn>=1.6.1",
]
